phycv-examples/requirements.txt
av==10.0.0
certifi==2022.12.7
charset-normalizer==3.0.1
contourpy==1.0.7
cycler==0.11.0
fonttools==4.38.0
idna==3.4
imageio==2.25.0
imageio-ffmpeg==0.4.8
kiwisolver==1.4.4
kornia==0.6.9
matplotlib==3.6.3
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
opencv-python==4.7.0.68
packaging==23.0
phycv==1.2.3
Pillow==9.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.28.2
six==1.16.0
torch==1.13.1
torchvision==0.14.1
typing_extensions==4.4.0
urllib3==1.26.14
#using python 3.10
